% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ccrop_zoning.R
\name{plot_ccrop_zoning}
\alias{plot_ccrop_zoning}
\title{Function to plot the climate crop zoning stack  generated in ccrop_zoning}
\usage{
plot_ccrop_zoning(zoning)
}
\arguments{
\item{zoning}{A stack generated in ccrop_zoning}
}
\value{
Returns a  plot (gg file) of monthly climate crop zoning.
}
\description{
This function will plot the ccrop_zoning.
}
\examples{
### Data preparation
\dontrun{
###ploting the cropzoning:

zoning_plot <- plot_ccrop_zoning(zoning)

### climatic zoning classes: 
                       - ST - suitable air temperature,
                       - RLT - Restricted due to low temperature,
                       - RHT - Restricted due to high temperature,
                       - ULT - Unsuitable due to low temperature,
                       - UHT - Unsuitable due to high temperature.
}
}
