% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_brazil_states.R
\name{see_brazil_states}
\alias{see_brazil_states}
\title{Function to see the Brazilian states available in cropZoning package and how we can use it in the loadROI function
\if{html}{\figure{logo_cropZoning.png}{options: height= 300 width=auto style = float:right alt= Our logo}}}
\usage{
see_brazil_states()
}
\value{
Returns a character strings listing the the Brazilian states available to run the cropZoning package.
}
\description{
This function will show the Brazilian state available in the package and how we can call each state polygon.
}
\examples{
\dontrun{

see_brazil_states()
}
}
