% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_tmax_and_tmin.R
\name{tmax_calibration}
\alias{tmax_calibration}
\title{Calibration of maximum air temperature from TerraClimate}
\usage{
tmax_calibration(slope, intercept, tmax_stack)
}
\arguments{
\item{slope}{the slope of the linear regression (numeric)}

\item{intercept}{the intercept of the linear regression (numeric)}

\item{tmax_stack}{stack of maximum air temperature}
}
\value{
Returns a Rasterstack of mean air temperature.
}
\description{
This fuction will calibrate the maximum air temperature from TerraClimate dataset based in the relationship with observed weather stations data.
}
\examples{
\dontrun{
tmax_cal<- tmax_calibrated(slope = 0.874787, intercept = 4.612894 , tmax_stack = tmax);
}
}
