% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmean.R
\name{tmean}
\alias{tmean}
\title{Mean air temperature calculation}
\usage{
tmean(tmax_stack, tmin_stack)
}
\arguments{
\item{tmax_stack}{stack of maximum air temperature}

\item{tmin_stack}{stack of minimum air temperature}
}
\value{
Returns a SpatRaster object of mean air temperature.
}
\description{
This fuction will calculate the mean air temperature from  minimum and maximum air temperature data.
}
\examples{
\dontrun{

image_tmean <- tmean(tmax_stack = image_tmax, tmin_stack = image_tmin)
}
}
