% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadROI.R
\name{loadROI}
\alias{loadROI}
\title{Load the data for the region of interest (ROI)}
\usage{
loadROI(variable, region, sub_region)
}
\arguments{
\item{variable}{Stack of minimum (tmin) or maximum (tmax) air temperature (Rasterstack).}

\item{region}{Use the "brazil" shp file to extract the rasterstack (variable) for one state (Brazilian state), or use the "biomes_brazil" to extract the Rasterstack (variable) for one biome of Brazil.}

\item{sub_region}{You have two options in this section, if you choice the brazil (in region parameter) you need to choice the Brazilian states, but if you choice the biomes-brasil (in region parameter) you must choice one of Brazilian biomes.}
}
\value{
Load the tmin (minimum air temperature) or tmax (maximum air temperature) rasterstack
}
\description{
This fuction will load the air temperature data for the region of interest (ROI).
}
\examples{
\dontrun{
#For Brazilian states

see_brazil_states()
image_tmin<-loadROI(variable = "tmin", region = "brazil", sub_region = 13) # sub_regions 1:27
image_tmax<-loadROI(variable = "tmax", region = "brazil", sub_region = 13)

#For Brazilian Biomes:

see_brazil_biomes()
image_tmin<- loadROI(variable = tmin, region = "biomes_brazil", sub_region = 2)# sub regions: 1:6 (biomes)

}
}
