% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ccrop_zoning.R
\name{plot_ccrop_zoning}
\alias{plot_ccrop_zoning}
\title{Function to plot the climate crop zoning stack  generated in ccrop_zoning or sr_ccrop_zoning}
\usage{
plot_ccrop_zoning(zoning_stack)
}
\arguments{
\item{zoning_stack}{stack}
}
\value{
Returns a  plot (gg file) of monthly climate crop zoning
}
\description{
This function will plot the ccrop_zoning map with the folowing classes: suitable; cold restriced; heat restriced, unsuitable-cold, and unsuitable-heat.
}
